+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Minor: Add setting to adjust the base network timeout seconds. (#904)
 - Minor: Allow routing Dink notifications through a HTTPS proxy server. (#859)
 - Bugfix: Updated PoH regions after their locations were moved in-game. (#903)
+- Dev: Send GIM shared bank contents to the custom metadata webhook upon opening the storage interface. (#895)
 
 ## 1.12.1